Add stricter host build warnings to match microVorbis (#29)
Add -Wpedantic, -Wconversion, -Wsign-conversion, -Wdouble-promotion, and
-Wformat=2 to the first-party wrapper flags in cmake/host.cmake and the
mp3_to_wav example. Upstream OpenCore sources stay suppressed with -w.

Fi sign-conversion in the host example.

+++ host_examples/mp3_to_wav/mp3_to_wav.cpp
@@ -219,7 +219,7 @@ int main(int argc, char* argv[]) {
                 break;
             }
 
-            stats.total_bytes_read += bytes_read;
+            stats.total_bytes_read += static_cast<size_t>(bytes_read);
 
             if (!decode_chunk(decoder, input_buffer.data(), static_cast<size_t>(bytes_read),
                               pcm_buffer, wav_writer, output_file, stats)) {
